spark-jupyter-workspace/finalytics_spark/pyproject.toml
[tool.poetry]
name = "finalytics-spark"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "finalytics_spark", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.5"
yfinance = "^0.2.51"
pyspark = "^3.5.4"
pyyaml = "^6.0.2"
jupyterlab-lsp = "^5.1.0"
nbimporter = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
